+++ ocrd_utils/ocrd_utils/constants.py
@@ -10,6 +10,8 @@ __all__ = [
     'MIME_TO_EXT',
     'PIL_TO_MIME',
     'MIME_TO_PIL',
+    'LOG_FORMAT',
+    'LOG_DATEFMT',
 ]
 
 VERSION = get_distribution('ocrd_utils').version
@@ -73,3 +75,7 @@ MIME_TO_PIL = {
     'image/x-portable-pixmap': 'PPM',
     'image/tiff': 'TIFF',
 }
+
+# Log level format implementing https://ocr-d.de/en/spec/cli#logging
+LOG_FORMAT = r'%(asctime)s.%(msecs)03d %(levelname)s %(name)s - %(message)s'
+LOG_DATEFMT = r'%Y-%m-%d %H:%M:%S'